Create and save the directory instance on disk
The current design is not good and required to be changed:
The size of dir_t on chisai-fs should be restricted to single block
size (now we assume a block should always be larger than 512 bytes).
The bad design limits the numbers of file in each directory and also limits
the name of files. The only advantage of this is the low complexity
of filesystem.