Add simple getattr test
The real purpose of this patch is to provide an extendable test
framework. We want to avoid damaging the original intact functions
when adding new features to the file system. With the framework, we can
simply add new test for the new features and also check that we can still
pass the operations which we could pass previously.