Fix the test case error of rewriting file
1. An error of find_first_set implementation of bitvec is fixed to avoid
allocation of improper block which is already in used

2. We add a 'REOPEN' which is a workaround for problem of truncate operation.
The truncate operation will reopen a file without syncing the same file that was
opened and not yet closed. When the flag is raised, any file operations that depend
on opened file should close it and reopen again.

3. For change 2, although the content of inode is changed by truncate operation, but
the inode number is still the same. So it could be more effective to find
inode from its index instead of using the file path. Our approach now is
easier for reading but has bad performance.